Split cover.py
